requirements.txt
PyQt5
openocd
pylink-square
